+++ pages/signin.tsx
@@ -1,54 +1,75 @@
+import {
+  INPUT_NAME,
+  SIGN_INPUT_ERROR_MESSAGES,
+} from "@/src/constant/SIGN_INPUT_TEXTS";
+import SignInPresenter from "@/src/containers/signin-page/signin.presenter";
+import signBlurError from "@/src/utils/sign-blur-error-message/signBlurError";
+import axios from "axios";
 import router from "next/router";
-import * as S from "@/styles/pages/sign-page.style";
+import { FormEvent, RefObject, useRef, useState } from "react";
 
-// Components
-import SignUpPageContainer from "@/src/containers/signin-page/useSignInPage";
-import SignPageHeader from "@/src/components/layout/sign-page/header/SignPageHeader";
-import SignInput from "@/src/components/sign-input/SignInput";
-import SocialSign from "@/src/components/layout/sign-page/social/SocialSign";
-import { useEffect } from "react";
+/**
+ * @description signupPage의 전반적인 로직이 담겨있습니다.
+ * @returns
+ */
+export default function SignIn() {
+  const [emailError, setEmailError] = useState("");
+  const [passwordError, setPasswordError] = useState("");
+  const emailRef = useRef<HTMLInputElement>(null);
+  const passwordRef = useRef<HTMLInputElement>(null);
 
-export default function SignUp() {
-  const {
+  const blurEvent = (type: string, ref: RefObject<HTMLInputElement>) => {
+    const userInput = ref.current!.value;
+    let setFunction;
+
+    switch (type) {
+      case "email":
+        setFunction = setEmailError;
+        break;
+
+      case "password-login":
+        setFunction = setPasswordError;
+        break;
+
+      default:
+        console.error("input의 타입 지정이 올바르지 않음");
+        return;
+    }
+    signBlurError(userInput, type, setFunction);
+  };
+
+  const submitEvent = async (e: FormEvent) => {
+    e.preventDefault();
+    const emailInput = emailRef.current!.value;
+    const passwordInput = passwordRef.current!.value;
+
+    try {
+      const {
+        data,
+      }: { data: { data: { accessToken: string; refreshToken: string } } } =
+        await axios.post("https://bootcamp-api.codeit.kr/api/sign-in", {
+          email: emailInput,
+          password: passwordInput,
+        });
+      localStorage.setItem("accessToken", data.data.accessToken);
+      localStorage.setItem("refreshToken", data.data.refreshToken);
+      router.push("/folder");
+    } catch {
+      signBlurError(emailInput, "email", setEmailError);
+      signBlurError(passwordInput, "password-login", setPasswordError);
+      setEmailError(SIGN_INPUT_ERROR_MESSAGES.NOT_CORRECT_EMAIL);
+      setPasswordError(SIGN_INPUT_ERROR_MESSAGES.NOT_CORRECT_PASSWORD);
+    }
+  };
+
+  const props = {
     emailError,
     passwordError,
     emailRef,
     passwordRef,
     blurEvent,
     submitEvent,
-  } = SignUpPageContainer();
+  };
 
-  useEffect(() => {
-    if (localStorage.getItem("accessToken")) {
-      router.push("/folder");
-    }
-  });
-
-  return (
-    <>
-      <S.SignPageWrapper>
-        <SignPageHeader
-          text="회원이 아니신가요?"
-          linkText="회원 가입하기"
-          linkTo="/signup"
-        />
-        <S.SignFormWrapper id="sign" onSubmit={(e) => submitEvent(e)}>
-          <SignInput
-            inputRef={emailRef}
-            inputType="email"
-            errorMessage={emailError}
-            blurEvent={() => blurEvent("email", emailRef)}
-          />
-          <SignInput
-            inputRef={passwordRef}
-            inputType="password-login"
-            errorMessage={passwordError}
-            blurEvent={() => blurEvent("password-login", passwordRef)}
-          />
-        </S.SignFormWrapper>
-        <S.SignSubmitButton form="sign">로그인</S.SignSubmitButton>
-        <SocialSign />
-      </S.SignPageWrapper>
-    </>
-  );
+  return <SignInPresenter props={props} />;
 }
